requirements.txt
multi-elo==2.0.0
mysql-connector-python==8.0.23
numpy==1.19.4
